final commit

+++ products/views.py
@@ -40,10 +40,10 @@ def product_detail_view(request,product_id):
     
     MRP = phone.price
     if phone.brand == 'Apple':
-        MRP = phone.price * 1.3
+        MRP = int(phone.price * 1.3)
     elif phone.brand == 'Samsung':
-        MRP = phone.price * 1.2
+        MRP = int(phone.price * 1.2)
     else:
-        MRP = phone.price * 1.1
+        MRP = int(phone.price * 1.1)
 
     return render(request,'products/product_details.html',{'phone':phone, 'MRP':MRP})
